requirements.txt
numpy==1.15.4
pytorch==1.4.0
torchvision==0.5.0
gym==0.17.1
mujoco_py==2.0.2.9
matplotlib==3.2.0
tqdm==4.46.0
scipy==1.4.1
tensorboardX==2.1